Added method stubs for legal moves
